Removed plan dependence on element class

+++ src/main_cuda.cpp
@@ -45,8 +45,8 @@ int main (int argc, char *argv[])
 	
 	
 	int n = inputParams ["gridpoints"].asInt / n_elements;
-	float position_0 = -1.0 + 2.0 / n_elements * id;
-	float position_n = -1.0 + 2.0 / n_elements * (id + 1);
+	double position_0 = -1.0 + 2.0 / n_elements * id;
+	double position_n = -1.0 + 2.0 / n_elements * (id + 1);
 	int excess_0;
 	int excess_n;
 	if (id == 0) {
@@ -70,7 +70,7 @@ int main (int argc, char *argv[])
 		process_messenger.add_boundary (one_d::edge_n, id + 1);
 	}
 	
-	cuda::one_d::chebyshev::fft_element <float> element (n, excess_0, position_0, excess_n, position_n, name, inputParams, &process_messenger, 0x00);
+	cuda::one_d::chebyshev::fft_element <double> element (n, excess_0, position_0, excess_n, position_n, name, inputParams, &process_messenger, 0x00);
 	
 	element.setup ();
